requirements.txt
# Apache Airflow ETL Pipeline Requirements
# These packages are needed to run the DAGs

# Core dependencies (already in base Airflow image)
apache-airflow==2.8.1
apache-airflow-providers-postgres

# Data processing
# Note: Using pandas 2.0.3 (compatible with Python 3.8 in Airflow 2.8.1)
pandas==2.0.3

# HTTP requests for API calls
requests==2.31.0

# Optional: Uncomment if you need these for your custom DAGs
# Note: Check Python version compatibility (Airflow 2.8.1 uses Python 3.8)
# numpy==1.24.4
matplotlib==3.7.3
# seaborn==0.12.2
# scipy==1.10.1
# scikit-learn==1.3.2
# plotly==5.18.0
sqlalchemy
streamlit

